pyproject.toml
[project]
name = "unifi-mcp"
version = "0.1.0"
description = "UniFi MCP servers for cloud and local controller management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1", 
    "fastmcp>=2.11.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "unifi-controller-api>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0", 
    "ruff>=0.12.7",
    "mypy>=1.13.0",
]

[project.scripts]
unifi-cloud-mcp = "unifi_mcp.cloud_server:main"
unifi-local-mcp = "unifi_mcp.local_server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
